Fix resource leak - too many YahooConnectors
Every single StoxSidebarItem, and at so many
opportunities for loops there was another provider
that spawned. This lead to lots of "too many files open".

The real solution is only having one provider for
the entire app, but this will stop an absurd amount
of threads from starting.

Also: don't infinite loop, no matter how glorious
it can be