Fix race condition in sidebar
Previously, if you clicked a sidebar item before the previous one had
finished loading, it the stock ticker and the company name/stock price
would not match in the sidebar. This could happen on a slow internet
connection or when pressing the up and down arrows at the same time.