Fix race-condition with graph color
When construct_graph calls the negative market change, it can call
it while the info is still refreshing ("---") or whatever placeholder
is showing. This can lead to incorrect statements depending on
the poll.

So lock the graph construction with the data poll mutex. While
at it, pass along extended info that could be helpful.

This does have a performance impact but hopefully that will be fixed.