Only allow a single window to be opened
GtkApplication tries to be smart when running the application multiple
times by just adding windows to one instance of the app. However, this
doesn't work very well due to the way Stox is currently architected; we
use things like global mutexes that are shared between windows. This led
to unexpected behavior like being unable to switch stocks due to data
being loaded in a completely different window. For now, only allow a
single window to be opened to fix this.