Use a list item factory instead of a set of ListBoxes
This has a set of advantages. If there is a lot of saved symbols a lot of
resources will be used, this ensures only what's needed is visible.

This might make it harder or easier for reordering, filtering and more advanced
things with the sidebar.

See https://gtk-rs.org/gtk4-rs/stable/latest/book/list_widgets.html for more